[caffe2] update the BlobSerializer acceptor to allow moving in the data
Update the `BlobSerializerBase` API so that the serizialized blob data is
passed as a `std::string&&` rather than `const std::string&`.  This allows the
acceptor to take ownership of the string data.  This allows the acceptor to do
things like queue it for storing asynchronously, rather than having to make a
copy of the data if they need it to remain valid after returning.

All existing `BlobSerializerBase` implementations already pass in a valid
rvalue reference to the data, so this change did not require updating any of
the existing serializer implementations.

Differential Revision: [D29204426](https://our.internmc.facebook.com/intern/diff/D29204426/)

ghstack-source-id: 131743003
Pull Request resolved: https://github.com/pytorch/pytorch/pull/60207